consolidate content scripts
